# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server golang pizza order service",
        "title": "Pizza Code Sample",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/orders": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "Creates an order",
                "parameters": [
                    {
                        "description": "Create order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ValidationError"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/orders/{orderID}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "Get order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "orderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ValidationError"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/pizzas": {
            "get": {
                "description": "Fetches the list of pizzas that can be ordered",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "Get list of pizzas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Pizza"
                            }
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.Order": {
            "type": "object",
            "required": [
                "name",
                "orderItems"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "Jimmy"
                },
                "orderItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.OrderItem"
                    }
                }
            }
        },
        "entities.OrderItem": {
            "type": "object",
            "required": [
                "pizzaName",
                "size"
            ],
            "properties": {
                "pizzaName": {
                    "type": "string",
                    "example": "Meatfeast"
                },
                "size": {
                    "type": "string",
                    "enum": [
                        "Large",
                        "Medium",
                        "Small"
                    ],
                    "example": "Large"
                }
            }
        },
        "entities.Pizza": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Meat",
                        "Cheese",
                        "Sauce"
                    ]
                },
                "name": {
                    "type": "string",
                    "example": "Meatfeast"
                },
                "price": {
                    "type": "number",
                    "example": 14.5
                }
            }
        },
        "errors.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}